// src/types/pools.ts
export type PoolFilterOptions = Ref<{
  poolIds?: string[];
  tokens?: string[];
  poolTypes?: PoolType[];
  useExactTokens?: boolean;
  pageSize?: number;
  first?: number;
  sortField?: string;
}>;